requirements.txt
chromadb==0.4.15
langchain-openai==0.0.8
streamlit==1.28.0
scikit-learn==1.2.2
nltk==3.8.1
psycopg2-binary==2.9.9